Use mTLS auth between flink and Kafka and introduce install scripts
1. The Kafka cluster runs a user operator
2. We install a user with mtls auth
3. We create another external nodeport listener with TLS auth enabled
4. We wire up the Flink deployment with the cluster ca certs, and the user
certs and key via environment variables. The job reads them and configures
the kafka clients.

Why:
We want to understand how the flink job can be configured to work with
Strimzi auth modes like mtls and scramsha.

The scripts should avoid copying commands out of the README. They are
split into:

- 01-setup.sh, delete and create minikube and install operators on it
- 02-flink.sh, install kafka, build the flink job image and deploy the flink job

Then it's on users to produce the transactions and consume alerts to see it
in action.